LayoutTests/fast/scrolling/scroll-animator-overlay-scrollbars-hovered-expected.txt

CONSOLE MESSAGE: MainLocalFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: MainLocalFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: LocalFrameView: mouseEnteredHorizontalScrollbar
CONSOLE MESSAGE: LocalFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: LocalFrameView: mouseExitedHorizontalScrollbar
CONSOLE MESSAGE: LocalFrameView: mouseExitedContentArea
CONSOLE MESSAGE: MainLocalFrameView: mouseMovedInContentArea
Test for https://bugs.webkit.org/show_bug.cgi?id=153304.


PASS successfullyParsed is true

TEST COMPLETE
